Merge PR instead of checking out
This commit switches from checking out the PR to merging it. When the
repo is cloned master is checked out, if the PR is then checked out, we
get master at the point the developer started working on the PR(or
latest rebase), not current master.

So instead we should assume that the current branch is master (as the
repo has just been checked out) and merge the PR into it, guaranteeing
that the PR is always tested agains't current master.