Remove custom ansible
A custom version of ansible is used to work around issues with SSH and
apt. This means that a different version of ansible is being used for
testing than is being used by deployments.

This commit removes the custom version of ansible. The SSH connection
issues should be addressed by https://review.openstack.org/#/c/152936/15
and the apt tasks are being modified so that they use do-until loops.