Fix tagging in commit-multinode.yml
There are several tasks that are not running when the prepare tag is
used because either the tag is missing or they are incorrectly tagged.

The syntax for tagging the included playbooks is incorrect, this commit
adjusts it so that they are run when the prepare tag is used.

This commit also adds the prepare tag to the task creating the vgs.