Remove use of cidr.py
cidr.py is used to determine the IP addresses to be assigned to certain
vars. The tasks that use the script run it on the box running ansible.
The pip dependencies that are required by the script are being
installed on the wrong hosts. The complexity that this script brings is
not required by commit-multinode.yml because the IP addresses are
constructed based on a member_number var set in the inventory.

This commit removes cidr.py and the associated code and replaces it
with the use of the member_numbar var format already used by the
commit-multinode.yml vars.