Set the tempest_public_subnet_cidr appropriately
We already have a br-vlan, we should set the tempest_public_subnet_cidr
to utilise this IP range.
Additionally, we can clean up some of the "used_ips" in the
commit_multinode vars, and adjust a typo on the storage_netowrk which
was set to use the same range as br-vlan.