Remove branch-vars
Theres no need to have vars files for each branch when jenkins-rpc
itself now has a branch to match each os-ansible-deployment branch.

This patch removes branch vars, and expects the branch/tag to build to
be passed to the playbooks via -e as os_ansible_branch. This can be a
branch, tag or SHA (commitish).

This will require jenkins jobs to be modified to pass the
OS_ANSIBLE_BRANCH parameter to docker, which will pass it to ansible as
os_ansible_branch.