Set default for OS_ANSIBLE_BRANCH
Commit bf55e25a3b4563d2caf882cf88e7608fc6e8b1b3 reworked how the branch
is specified so that it is now defined by the variable
OS_ANSIBLE_BRANCH. No default value is set for this variable. The
previously used variable of TARGET_BRANCH has also been left in the
commit-multinode.sh script even though it no longer serves any purpose.

This commit sets a default value of master for OS_ANSIBLE_BRANCH and
removes TARGET_BRANCH.