Move ansible.cfg so that automatically used
The current location of ansible.cfg in the repo means that it is not
automatically used when running the playbooks.

This commit moves ansible.cfg into the playbooks directory so that it
is automatically used when the playbooks are run.

(cherry picked from commit d752fad6851ec2b9a1c3d7650437085dbed91dc7)