Create interface config for commit cluster em1
The default config for em1 is overwritten when the network role runs,
and its config is not replaced. This means that em1 will not have an ip
if any boxes are rebooted.

This is not normally a problem as they are rekicked most times, but can
be a pain when troubleshooting.