Add upgrade tag
Adds an upgrade function to commit-multinode. This is a simple function
that will curl a script and run it on the first infra node. The
intention is to test in development upgrade scripts.

Also ensures all tags are run regardless of exit status, so cleanup can
be a tag.

(cherry picked from commit 68e1bac833c81e9a80b55e67a76b13d87b0eb30d)