Dont accumulate return codes
Currently the final return code is the sum of previous return codes.
This can be problematic as if the return codes sum to 256, that will be
intepreted as success when it actually represents multiple failures.

This patch ensures the return code is 0 for success or 1 for failure
regardless of the number of failures.

(cherry picked from commit f9a2d00db90e4aa34d8fbbc94f31fa0a3a5ca694)